+++ components/director/internal/domain/systemauth/repository.go
@@ -43,11 +43,11 @@ type repository struct {
 func NewRepository(conv Converter) *repository {
 	return &repository{
 		creator:            repo.NewCreatorGlobal(resource.SystemAuth, tableName, tableColumns),
-		singleGetter:       repo.NewSingleGetterWithEmbeddedTenant(resource.SystemAuth, tableName, tenantColumn, tableColumns),
+		singleGetter:       repo.NewSingleGetterWithEmbeddedTenant(tableName, tenantColumn, tableColumns),
 		singleGetterGlobal: repo.NewSingleGetterGlobal(resource.SystemAuth, tableName, tableColumns),
-		lister:             repo.NewListerWithEmbeddedTenant(resource.SystemAuth, tableName, tenantColumn, tableColumns),
+		lister:             repo.NewListerWithEmbeddedTenant(tableName, tenantColumn, tableColumns),
 		listerGlobal:       repo.NewListerGlobal(resource.SystemAuth, tableName, tableColumns),
-		deleter:            repo.NewDeleterWithEmbeddedTenant(resource.SystemAuth, tableName, tenantColumn),
+		deleter:            repo.NewDeleterWithEmbeddedTenant(tableName, tenantColumn),
 		deleterGlobal:      repo.NewDeleterGlobal(resource.SystemAuth, tableName),
 		updater:            repo.NewUpdaterWithEmbeddedTenant(resource.SystemAuth,tableName,[]string{"value"},tenantColumn,[]string{"id"}),
 		conv:               conv,
@@ -68,7 +68,7 @@ func (r *repository) Create(ctx context.Context, item model.SystemAuth) error {
 // GetByID missing godoc
 func (r *repository) GetByID(ctx context.Context, tenant, id string) (*model.SystemAuth, error) {
 	var entity Entity
-	if err := r.singleGetter.Get(ctx, tenant, repo.Conditions{repo.NewEqualCondition("id", id)}, repo.NoOrderBy, &entity); err != nil {
+	if err := r.singleGetter.Get(ctx, resource.SystemAuth, tenant, repo.Conditions{repo.NewEqualCondition("id", id)}, repo.NoOrderBy, &entity); err != nil {
 		return nil, err
 	}
 
@@ -127,7 +127,7 @@ func (r *repository) ListForObject(ctx context.Context, tenant string, objectTyp
 		repo.NewEqualCondition(objTypeFieldName, objectID),
 	}
 
-	err = r.lister.List(ctx, tenant, &entities, conditions...)
+	err = r.lister.List(ctx, resource.SystemAuth, tenant, &entities, conditions...)
 
 	if err != nil {
 		return nil, err
@@ -192,7 +192,7 @@ func (r *repository) DeleteAllForObject(ctx context.Context, tenant string, obje
 	if objectType == model.IntegrationSystemReference {
 		return r.deleterGlobal.DeleteManyGlobal(ctx, repo.Conditions{repo.NewEqualCondition(objTypeFieldName, objectID)})
 	}
-	return r.deleter.DeleteMany(ctx, tenant, repo.Conditions{repo.NewEqualCondition(objTypeFieldName, objectID)})
+	return r.deleter.DeleteMany(ctx, resource.SystemAuth, tenant, repo.Conditions{repo.NewEqualCondition(objTypeFieldName, objectID)})
 }
 
 // DeleteByIDForObject missing godoc
@@ -205,7 +205,7 @@ func (r *repository) DeleteByIDForObject(ctx context.Context, tenant, id string,
 	}
 	objTypeCond = repo.NewNotNullCondition(column)
 
-	return r.deleter.DeleteOne(ctx, tenant, repo.Conditions{repo.NewEqualCondition("id", id), objTypeCond})
+	return r.deleter.DeleteOne(ctx, resource.SystemAuth, tenant, repo.Conditions{repo.NewEqualCondition("id", id), objTypeCond})
 }
 
 // DeleteByIDForObjectGlobal missing godoc
